pyproject.toml
[tool.poetry]
name = "api-for-statistics"
version = "0.1.0"
description = "API для просмотра статистики по приложению \"Я здесь живу\""
authors = ["Michael <pozdinmisha666@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
clickhouse-sqlalchemy = "0.2.4"
asyncio = "^3.4.3"
uvicorn = {extras = ["standart"], version = "^0.30.1"}
gunicorn = "^22.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
